Others: 将 mod 函数重命名为 int_mod

+++ others/int-ecc/int-ecc.h
@@ -17,7 +17,7 @@ struct param {
 };
 
 int ecc_point_add(int p, int a, const struct point *p1, const struct point *p2, struct point *p3);
-int ecc_point_mul(int p, int a, unsigned long x, const struct point *p1, struct point *p2);
+int ecc_point_mul(int p, int a, int x, const struct point *p1, struct point *p2);
 
 int ecc_point_on_curve(int p, int a, int b, const struct point *p1);
